Rename RKernel SDK submodules
