+++ enclave_apps/oak_echo_enclave_app/src/main.rs
@@ -23,9 +23,10 @@ extern crate alloc;
 use alloc::boxed::Box;
 
 use oak_restricted_kernel_sdk::{
+    attestation::InstanceEvidenceProvider,
+    crypto::InstanceEncryptionKeyHandle,
     channel::{start_blocking_server, FileDescriptorChannel},
     entrypoint,
-    instance_attestation::InstanceEvidenceProvider,
     utils::samplestore::StaticSampleStore,
 };
