+++ enclave_apps/oak_functions_enclave_app/src/main.rs
@@ -24,9 +24,10 @@ use alloc::{boxed::Box, sync::Arc};
 
 use oak_functions_service::wasm::WasmHandler;
 use oak_restricted_kernel_sdk::{
+    attestation::InstanceEvidenceProvider,
+    crypto::InstanceEncryptionKeyHandle,
     channel::{start_blocking_server, FileDescriptorChannel},
     entrypoint,
-    instance_attestation::{InstanceEncryptionKeyHandle, InstanceEvidenceProvider},
     utils::samplestore::StaticSampleStore,
 };
