+++ testing/oak_echo_service/Cargo.toml
@@ -20,5 +20,5 @@ micro_rpc_build = { workspace = true }
 [dev-dependencies]
 async-trait = "*"
 oak_restricted_kernel_sdk = { workspace = true, features = [
-  "mock_attestation",
+  "mock_sdk",
 ] }
