+++ testing/oak_echo_service/tests/integration_test.rs
@@ -30,7 +30,7 @@ const TEST_DATA: &[u8] = b"test_data";
 #[test]
 fn it_should_handle_echo_requests() {
     let evidence_provider =
-        oak_restricted_kernel_sdk::mock_attestation::MockEvidenceProvider::create()
+        oak_restricted_kernel_sdk::mock_sdk::MockEvidenceProvider::create()
             .expect("couldn't create evidence provider");
     let service = EchoService { evidence_provider };
     let mut client = EchoClient::new(EchoServer::new(service));
@@ -45,7 +45,7 @@ fn it_should_handle_echo_requests() {
 #[test]
 fn it_should_provide_evidence() {
     let evidence_provider =
-        oak_restricted_kernel_sdk::mock_attestation::MockEvidenceProvider::create()
+        oak_restricted_kernel_sdk::mock_sdk::MockEvidenceProvider::create()
             .expect("couldn't create evidence provider");
     let service = EchoService { evidence_provider };
     let mut client = EchoClient::new(EchoServer::new(service));
